undo renaming mags
