+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/dataformats/n5/N5BucketProvider.scala
@@ -25,7 +25,7 @@ class N5BucketProvider(layer: N5Layer,
   override def openDatasetArrayHandle(readInstruction: DataReadInstruction)(
       implicit ec: ExecutionContext): Fox[DatasetArray] = {
     val magLocatorOpt: Option[MagLocator] =
-      layer.magLocators.find(_.mag == readInstruction.bucket.mag)
+      layer.mags.find(_.mag == readInstruction.bucket.mag)
 
     magLocatorOpt match {
       case None => Fox.empty
