+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/dataformats/n5/N5DataLayers.scala
@@ -18,7 +18,7 @@ trait N5Layer extends DataLayerWithMagLocators {
                      sharedChunkContentsCache: Option[AlfuCache[String, MultiArray]]) =
     new N5BucketProvider(this, dataSourceId, remoteSourceDescriptorServiceOpt, sharedChunkContentsCache)
 
-  def resolutions: List[Vec3Int] = magLocators.map(_.mag)
+  def resolutions: List[Vec3Int] = mags.map(_.mag)
 
   def lengthOfUnderlyingCubes(resolution: Vec3Int): Int = Int.MaxValue // Prevents the wkw-shard-specific handle caching
 
@@ -30,7 +30,7 @@ case class N5DataLayer(
                         category: Category.Value,
                         boundingBox: BoundingBox,
                         elementClass: ElementClass.Value,
-                        magLocators: List[MagLocator],
+                        mags: List[MagLocator],
                         defaultViewConfiguration: Option[LayerViewConfiguration] = None,
                         adminViewConfiguration: Option[LayerViewConfiguration] = None,
                         coordinateTransformations: Option[List[CoordinateTransformation]] = None,
@@ -46,7 +46,7 @@ case class N5SegmentationLayer(
                                 name: String,
                                 boundingBox: BoundingBox,
                                 elementClass: ElementClass.Value,
-                                magLocators: List[MagLocator],
+                                mags: List[MagLocator],
                                 largestSegmentId: Option[Long],
                                 mappings: Option[Set[String]] = None,
                                 defaultViewConfiguration: Option[LayerViewConfiguration] = None,
