+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/dataformats/precomputed/PrecomputedDataLayers.scala
@@ -27,7 +27,7 @@ trait PrecomputedLayer extends DataLayerWithMagLocators {
                      sharedChunkContentsCache: Option[AlfuCache[String, MultiArray]]) =
     new PrecomputedBucketProvider(this, dataSourceId, remoteSourceDescriptorServiceOpt, sharedChunkContentsCache)
 
-  def resolutions: List[Vec3Int] = magLocators.map(_.mag)
+  def resolutions: List[Vec3Int] = mags.map(_.mag)
 
   def lengthOfUnderlyingCubes(resolution: Vec3Int): Int = Int.MaxValue // Prevents the wkw-shard-specific handle caching
 
@@ -39,7 +39,7 @@ case class PrecomputedDataLayer(
                                  boundingBox: BoundingBox,
                                  category: Category.Value,
                                  elementClass: ElementClass.Value,
-                                 magLocators: List[MagLocator],
+                                 mags: List[MagLocator],
                                  defaultViewConfiguration: Option[LayerViewConfiguration] = None,
                                  adminViewConfiguration: Option[LayerViewConfiguration] = None,
                                  coordinateTransformations: Option[List[CoordinateTransformation]] = None,
@@ -55,7 +55,7 @@ case class PrecomputedSegmentationLayer(
                                          name: String,
                                          boundingBox: BoundingBox,
                                          elementClass: ElementClass.Value,
-                                         magLocators: List[MagLocator],
+                                         mags: List[MagLocator],
                                          largestSegmentId: Option[Long],
                                          mappings: Option[Set[String]] = None,
                                          defaultViewConfiguration: Option[LayerViewConfiguration] = None,
